backend/requirements.txt
# Backend dependencies
fastapi
uvicorn[standard]
pydantic
python-dotenv
google-generativeai